backend/requirements.txt
# Fake News Verification Backend

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.3.0
httpx>=0.26.0
pydantic>=2.0.0
python-dotenv>=1.0.0
